requirements.txt
arrow==0.14.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
icalendar==4.0.7
ics==0.7
idna==2.10
lxml==4.5.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
TatSu==5.5.0
urllib3==1.25.10
